+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 * `wn.Lexicon.specifier()`
-
+* `wn.config.allow_multithreading` ([#86])
 
 ### Removed
 
@@ -192,6 +192,7 @@ abandoned, but this is an entirely new codebase.
 [#79]: https://github.com/goodmami/wn/issues/79
 [#81]: https://github.com/goodmami/wn/issues/81
 [#83]: https://github.com/goodmami/wn/issues/83
+[#86]: https://github.com/goodmami/wn/issues/86
 [#90]: https://github.com/goodmami/wn/issues/90
 [#92]: https://github.com/goodmami/wn/issues/92
 [#93]: https://github.com/goodmami/wn/issues/93